Switch to bigDecimal

+++ src/main/java/com/adrians/rates/entity/FxRateEntity.java
@@ -2,6 +2,7 @@ package com.adrians.rates.entity;
 
 import jakarta.persistence.*;
 
+import java.math.BigDecimal;
 import java.time.LocalDate;
 
 @Entity
@@ -13,12 +14,12 @@ public class FxRateEntity {
     private Long id;
 
     private String currency;
-    private double rate;
+    private BigDecimal rate;
     private LocalDate date;
 
     public FxRateEntity() {}
 
-    public FxRateEntity(String currency, double rate, LocalDate date) {
+    public FxRateEntity(String currency, BigDecimal rate, LocalDate date) {
         this.currency = currency;
         this.rate = rate;
         this.date = date;
@@ -26,10 +27,10 @@ public class FxRateEntity {
 
     public Long getId() { return id; }
     public String getCurrency() { return currency; }
-    public double getRate() { return rate; }
+    public BigDecimal getRate() { return rate; }
     public LocalDate getDate() { return date; }
 
     public void setCurrency(String currency) { this.currency = currency; }
-    public void setRate(double rate) { this.rate = rate; }
+    public void setRate(BigDecimal rate) { this.rate = rate; }
     public void setDate(LocalDate date) { this.date = date; }
 }
